pyproject.toml
[tool.poetry]
name = "keep-event-handler"
version = "0.1.0"
description = "Event Handler service for Keep"
authors = ["Keep Authors"]
readme = "README.md"
packages = [{include = "keep-event-handler"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
python-dotenv = "^0.21.1"
aiokafka = "^0.11.0"
redis = "^5.0.0"
arq = "^0.26.3"
pydantic = "^1.10.4"
prometheus-fastapi-instrumentator = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
